Refactor OutputField component and add getRandomAffirmation utility
