+++ src/components/OutputField.svelte
@@ -1,77 +1,71 @@
 <script lang="ts">
-	import { formatAssignmentFilename } from '$lib/formatAssignmentFilename';
-	import type { AssignmentInfo } from '$lib/types';
-	import { copyToClipboard } from '$lib/utils';
+  import { formatAssignmentFilename } from '$lib/formatAssignmentFilename';
+  import type { AssignmentInfo } from '$lib/types';
+  import { copyToClipboard } from '$lib/utils';
 
-	import { confetti } from '@neoconfetti/svelte';
-	import { tick } from 'svelte';
-	import CopyIcon from './CopyIcon.svelte';
-	import toast from 'svelte-french-toast';
-	import { affirmations } from '$lib/constants';
+  import { Confetti } from 'svelte-confetti';
 
-	type Props = {
-		values: AssignmentInfo;
-	};
+  import { getNextAffirmation } from '$lib/getRandomAffirmation';
+  import { tick } from 'svelte';
+  import toast from 'svelte-french-toast';
+  import CopyIcon from './CopyIcon.svelte';
 
-	let { values } = $props<Props>();
+  type Props = {
+    values: AssignmentInfo;
+  };
 
-	let formattedFileName = $derived(formatAssignmentFilename(values));
-	let confettiEnabled = $state(false);
+  let { values } = $props<Props>();
 
-	function shuffleArray(array: Array<string>) {
-		for (let i = array.length - 1; i > 0; i--) {
-			const j = Math.floor(Math.random() * (i + 1));
-			[array[i], array[j]] = [array[j], array[i]];
-		}
-		return array;
-	}
-	const shuffledAffirmations = shuffleArray(affirmations);
+  let formattedFileName = $derived(formatAssignmentFilename(values));
+  let confettiEnabled = $state(false);
 
-	let currentIndex = 0;
+  async function handleClick(formattedFileName: string) {
+    copyToClipboard(formattedFileName);
 
-	async function handleClick(formattedFileName: string) {
-		copyToClipboard(formattedFileName);
+    showConfetti();
 
-		showConfetti();
+    const affirmation = getNextAffirmation();
 
-		const affirmation = shuffledAffirmations[currentIndex % shuffledAffirmations.length];
-		currentIndex++;
+    toast.success(`Copied to clipboard.\n${affirmation}`, {
+      duration: 5000,
+      style:
+        'background: hsl(var(--card)); color: hsl(var(--card-foreground)); border: 1px solid hsl(var(--border));',
+    });
+  }
 
-		toast.success(`Copied to clipboard.\n${affirmation}`, {
-			duration: 5000,
-			style:
-				'background: hsl(var(--card)); color: hsl(var(--card-foreground)); border: 1px solid hsl(var(--border));'
-		});
-	}
-
-	async function showConfetti() {
-		confettiEnabled = false;
-		await tick();
-		confettiEnabled = true;
-	}
-
-	function getRandFloat(min: number, max: number) {
-		return Math.random() * (max - min) + min;
-	}
+  async function showConfetti() {
+    confettiEnabled = false;
+    await tick();
+    confettiEnabled = true;
+  }
 </script>
 
 <button
-	on:click={() => handleClick(formattedFileName)}
-	class="group relative flex w-full items-center justify-center gap-2 rounded-md border border-opacity-0 bg-inputBg p-2 px-8 text-inputBg-foreground shadow-inner hover:border-opacity-100 hover:brightness-110"
+  on:click={() => handleClick(formattedFileName)}
+  class="group relative flex w-full items-center justify-center gap-2 rounded-md border border-opacity-0 bg-inputBg p-2 px-8 text-inputBg-foreground shadow-inner hover:border-opacity-100 hover:brightness-110"
 >
-	<div class="relative">
-		{formattedFileName}
-		<div class="absolute right-1/2">
-			{#if confettiEnabled}
-				<div
-					use:confetti={{ particleSize: 4, particleCount: 250, force: getRandFloat(0.3, 0.9) }}
-				/>
-			{/if}
-		</div>
-	</div>
-	<div
-		class="absolute right-4 w-4 fill-foreground opacity-40 transition-opacity duration-300 group-hover:opacity-100"
-	>
-		<CopyIcon style="w-4" />
-	</div>
+  <div class="relative">
+    {formattedFileName}
+    <div class="absolute right-1/2">
+      {#if confettiEnabled}
+        <Confetti
+          delay={[0, 150]}
+          x={[-2, 2]}
+          y={[3, 0]}
+          duration={2200}
+          amount={250}
+          size={5}
+          xSpread={0.1}
+          rounded={true}
+          fallDistance={'100px'}
+        />
+      {/if}
+    </div>
+  </div>
+
+  <div
+    class="absolute right-4 w-4 fill-foreground opacity-40 transition-opacity duration-300 group-hover:opacity-100"
+  >
+    <CopyIcon style="w-4" />
+  </div>
 </button>
